+++ CHANGELOG.md
@@ -10,3 +10,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added reference to System.Web in `prefix.ps1` that imports assembly durning module import
 - Added reference to System.Net.Http in `prefix.ps1` that imports assembly durning module import
 - Added a new function to validate the api session `Test-ApiSession.ps1`
+
+- Updated control flow for setting up the API Session